ct: Move ct-partition to the ct/frontend
The 'frontend' is functionally equivalent to 'partition_proxy'
implementation (cloud_topic_partition) but it doesn't depend on kafka
layer and it lives in cloud_topics module.

Use 'frontend' to implement 'cloud_topic_partition'. This makes
'cloud_topic_partition' a wrapper around the 'frontend' and all cloud
topics related logic is contained inside the src/v/cloud_topics module.
