+++ src/v/cloud_topics/frontend/BUILD
@@ -9,9 +9,9 @@ redpanda_cc_library(
         "level_zero_reader_impl.h",
     ],
     implementation_deps = [
-        "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:app",
         "//src/v/cloud_topics:data_plane_api",
+        "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cluster",
         "//src/v/config",
@@ -23,6 +23,49 @@ redpanda_cc_library(
         "//src/v/cloud_topics:extent_meta",
         "//src/v/model",
         "//src/v/storage",
+        "@boost//:outcome",
+        "@seastar",
+    ],
+)
+
+redpanda_cc_library(
+    name = "errc",
+    srcs = [],
+    hdrs = [
+        "errc.h",
+    ],
+    include_prefix = "cloud_topics/frontend",
+    visibility = ["//visibility:public"],
+    deps = [],
+)
+
+redpanda_cc_library(
+    name = "frontend",
+    srcs = [
+        "frontend.cc",
+    ],
+    hdrs = [
+        "frontend.h",
+    ],
+    implementation_deps = [
+        "//src/v/cloud_topics:logger",
+        "//src/v/cluster",
+    ],
+    include_prefix = "cloud_topics/frontend",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":errc",
+        ":level_zero_reader_impl",
+        "//src/v/base",
+        "//src/v/cloud_storage",
+        "//src/v/cloud_topics:app",
+        "//src/v/cloud_topics:data_plane_api",
+        "//src/v/cloud_topics:extent_meta",
+        "//src/v/cloud_topics:placeholder",
+        "//src/v/model",
+        "//src/v/raft",
+        "//src/v/storage",
+        "//src/v/storage:record_batch_builder",
         "@boost//:outcome",
         "@seastar",
     ],
